python_backend/requirements.txt
fastapi
uvicorn
torch
pymongo[srv]
python-dotenv
pypdf
python-multipart
